+++ engine/src/main/scala/org/scalarules/dsl/nl/datum/DatumImplicits.scala
@@ -1,4 +1,4 @@
-package org.scalarules.dsl.nl.grammar.datum
+package org.scalarules.dsl.nl.datum
 
 import java.util.Date
