+++ engine/src/main/scala/org/scalarules/dsl/nl/datum/package.scala
@@ -1,0 +1,5 @@
+package org.scalarules.dsl.nl
+
+package object datum extends DatumImplicits {
+
+}
